pyproject.toml
[tool.poetry]
name = "seattle_map"
version = "0.1.0"
description = ""
authors = ["Reuben Britto <reubenbritto@ip-192-168-1-7.ec2.internal>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.1.1"
requests = "^2.28.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
